requirements.txt
discord.py==0.16.12
xkcd==2.4.0
